front end stuff, js and haml

+++ db/migrate/20150204203424_create_event_managers.rb
@@ -2,7 +2,7 @@ class CreateEventManagers < ActiveRecord::Migration
   def change
     create_table :event_managers do |t|
       t.decimal :stdev
-      t.integer :mean
+      t.decimal :mean
       t.integer :number_races
       t.string :event_name
       t.references :runner
